vercel-app/requirements.txt
google-generativeai==0.8.3
pandas==2.2.2
openpyxl==3.1.5
PyPDF2==3.0.1
requests==2.32.3